% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_halfcourt}
\alias{draw_halfcourt}
\title{Plot Half Court}
\usage{
draw_halfcourt(xlim = c(-300, 300), ylim = c(-100, 500), ...)
}
\value{
Plot of half court
}
\description{
Plot the court markings for the NBA regulation basketball half court.
}
\examples{
sc_data <- shots_cavs
# plot halfcout markings
draw_halfcourt()
# plot data
points(sc_data$LOC_X, sc_data$LOC_Y, cex = 0.5)
# highlight three-pointers
threes <- dplyr::filter(sc_data, SHOT_TYPE == "3PT Field Goal")
points(threes$LOC_X, threes$LOC_Y, cex = 0.5, col = "#FF6688")

}
