% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{game_id}
\alias{game_id}
\title{Get Game Identifiers for NBA}
\usage{
game_id(year = "", game_number = "")
}
\arguments{
\item{year}{Year the season starts in YYYY format.}

\item{game_number}{Game number in the season.}
}
\examples{

}
