% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_boxscoresummaryv2}
\alias{get_boxscoresummaryv2}
\title{Get Boxscore Summary (V2)}
\usage{
get_boxscoresummaryv2(GameID = "", ...)
}
\arguments{
\item{GameID}{See \code{\link[nba]{make_url}}.}
}
\value{
A data frame containing team/player/arena information for a given
  game.
}
\description{
Get Boxscore Summary (V2)
}
\examples{
get_boxscoresummaryv2(GameID = "0021300028")

}
