% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_pbp}
\alias{get_pbp}
\title{Get Play-by-Play Data}
\usage{
get_pbp(GameID = "", StartPeriod = "", EndPeriod = "", ...)
}
\arguments{
\item{GameID}{Game identifier.}

\item{StartPeriod}{TBD.}

\item{EndPeriod}{TBD.}

\item{...}{Arguments to code{\link[nba]{make_url}}.}
}
\value{
A data frame containing basic play-by-play information for a single game.
}
\description{
Get Play-by-Play Data
}
