% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_playerinfo}
\alias{get_playerinfo}
\title{Get Player Information Data}
\usage{
get_playerinfo(LeagueID = "", Season = "", IsOnlyCurrentSeason = "", ...)
}
\arguments{
\item{LeagueID}{Game identifier.}

\item{Season}{TBD.}

\item{IsOnlyCurrentSeason}{TBD.}

\item{...}{Arguments to {\link[nba]{make_url}}.}
}
\value{
A data frame containing player information in a given season.
}
\description{
Get Player Information Data
}
\examples{
# get player information for the 2016-17 NBA season
player_information <- get_playerinfo(LeagueID = "00",
                                     Season = "2016-17",
                                     IsOnlyCurrentSeason = "1")

}
