% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_teaminfo}
\alias{get_teaminfo}
\title{Get Team Year Availability Information}
\usage{
get_teaminfo(LeagueID = "", ...)
}
\arguments{
\item{LeagueID}{See \code{\link[nba]{make_url}}.}

\item{...}{Arguments to \code{\link[nba]{make_url}}.}
}
\value{
A data frame containing player information in a given season.
}
\description{
Get Team Year Availability Information
}
\examples{
# get team information for the NBA
team_roster <- get_teamroster(TeamID = "1610612748",
                              Season = "2016-17")

}
