% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_teamroster}
\alias{get_teamroster}
\title{Get Team Roster Information}
\usage{
get_teamroster(TeamID = "", Season = "", ...)
}
\arguments{
\item{TeamID}{See \code{\link[nba]{make_url}}.}

\item{Season}{See \code{\link[nba]{make_url}}.}

\item{...}{Arguments to \code{\link[nba]{make_url}}.}
}
\value{
A data frame containing player information in a given season.
}
\description{
Get Team Roster Information
}
\examples{
# get team roster information for Miami Heat in the 2016-17 NBA season
team_roster <- get_teamroster(TeamID = "1610612748",
                              Season = "2016-17")

}
