% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_url.R
\name{make_url}
\alias{make_url}
\title{Construct NBA API URL}
\usage{
make_url(datatype = NULL, SeasonType = "", LeagueID = "", Season = "",
  IsOnlyCurrentSeason = "", PlayerID = "", TeamID = "", GameID = "",
  ContextMeasure = "", PlayerPosition = "", DateFrom = "", DateTo = "",
  GameSegment = "", LastNGames = "", Location = "", Month = "",
  OpponentTeamID = "", Outcome = "", SeasonSegment = "",
  VSConference = "", VSDivision = "", RookieYear = "", Period = "",
  StartPeriod = "", EndPeriod = "", StartRange = "", EndRange = "",
  RangeType = "")
}
\arguments{
\item{datatype}{One of \code{"shotchartdetail"}, ...}

\item{SeasonType}{One of \code{"Regular+Season"}, \code{"Pre+Season"},\code{"Playoffs"}, or \code{"All+Star"}.}

\item{LeagueID}{One of \code{"00"} (NBA) or \code{"01"} (ABA).}

\item{Season}{Of the format \code{"YYYY-YY"}. For example \code{"2016-17"}.}

\item{IsOnlyCurrentSeason}{TBD.}

\item{PlayerID}{Player identifier.}

\item{TeamID}{Team identifier.}

\item{GameID}{Game identifier.}

\item{ContextMeasure}{TBD.}

\item{PlayerPosition}{TBD.}

\item{DateFrom}{TBD.}

\item{DateTo}{TBD.}

\item{GameSegment}{TBD.}

\item{LastNGames}{TBD.}

\item{Location}{TBD.}

\item{Month}{One of \code{"0"} (all months), \code{"1"} to \code{"7"} (regular season), or \code{"8"}/\code{"9"} (playoffs).}

\item{OpponentTeamID}{TBD.}

\item{Outcome}{TBD.}

\item{SeasonSegment}{TBD.}

\item{VSConference}{TBD.}

\item{VSDivision}{TBD.}

\item{RookieYear}{TBD.}

\item{Period}{TBD.}

\item{StartPeriod}{TBD.}

\item{EndPeriod}{TBD.}
}
\value{
NBA API url string.
}
\description{
Construct an NBA URL given various API arguments.
}
