% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{merge_shot_data}
\alias{merge_shot_data}
\title{Merge Shot Data with Raster Half Court}
\usage{
merge_shot_data(shpfile, shot_df)
}
\arguments{
\item{shpfile}{SpatialPolygonsDataFrame returned \code{from half_court_raster}.}

\item{shot_df}{Data frame object returned from \code{get_shotchart}.}
}
\value{
SpatialPolygonsDataFrame object which contains the spatial polygons
  along with additional data from(shot_df) in the data slot.
}
\description{
Merge the shot chart data with a SpatialPolygonDataFrame representing an NBA
regulation half court. This makes it useful to construct a weigth matrix that
can be used in spatial modeling.
}
