% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-package.R
\name{NBAapi-package}
\alias{NBAapi-package}
\alias{NBAapi}
\title{Accessing the NBA API}
\description{
The \pkg{NBAapi} package is designed to make accessing the NBA API
(\link{stat.com.nba}) easier. It provides functions to download, for example,
shot chart data or play-by-play data. Additionally it provides graphical
functions to plot court markings, which can be used in conjunction with the
shot chart data.
}
