% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{new_hex}
\alias{new_hex}
\title{Origin Hexgon}
\usage{
new_hex(hex_origin, coordinates, scale = 1)
}
\description{
Construct a hexagon that has the same area a hexagon polygon from
\code{\link[nba]{half_court_hex}} with center at the origin.
}
