% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{player_ids}
\alias{player_ids}
\title{Get Player Identifiers for NBA}
\usage{
player_ids(Season = "", IsOnlyCurrentSeason = "")
}
\arguments{
\item{Season}{See \code{\link[nba]{make_url}}.}

\item{IsOnlyCurrentSeason}{See \code{\link[nba]{make_url}}.}
}
\value{
A data frame returning player IDs, names, and team affiliations.
}
\description{
Wrapper for \code{\link[nba]{get_playerinfo}} to get all NBA player
identifiers.
}
\examples{
player_ids(Season = "2016-17", IsOnlyCurrentSeason = "0")

}
