% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_hexes}
\alias{plot_hexes}
\title{Plot Hexagon Shot Chart}
\usage{
plot_hexes(dat, player, ..., variable = "FGP", cellsize = 15,
  scale = TRUE)
}
\arguments{
\item{dat}{Data frame object returned from {\link[nba]{get_shotchart}}.}

\item{player}{Player name string or vector of player name strings.}

\item{...}{Additional arguments to pass to {\link[dplyr]{filter}}.}

\item{variable}{One of \code{"FGP"} or \code{"FGPvLeague"} to color code the
hexagons according to player field goal percentage or the difference
between player field goal percentage and the overall league, respectively.}

\item{cellsize}{Number of hexagon cells to construct along the x dimension. A
larger value results in a finer lattice.}

\item{scale}{Logical. If set to \code{TRUE} then the hexagons are scaled
according to shot frequency.}
}
\value{
Hexagons overlayed on a plot. Note that you need to run the function
  {\link[nba]{draw_halfcourt}} in order to plot the hexgons on the
  appropriate court.
}
\description{
Plot hexagon shot chart for a player or set of players.
}
\examples{
# plot shots by LeBron James in 2016-17
draw_halfcourt()
plot_hexes(shots_1617, player = "LeBron James", variable = "FGPvLeague",
           scale = T, cellsize = 15,
           LOC_Y <= 500)

# omit the restricted zone
draw_halfcourt()
plot_hexes(shots_1617, player = "LeBron James", variable = "FGPvLeague",
           LOC_Y <= 500, SHOT_ZONE_BASIC != "Restricted Area")


}
