% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resize_hexes}
\alias{resize_hexes}
\title{Resize Hexagons}
\usage{
resize_hexes(hex_grid)
}
\description{
Resize hexagons according to the frequency of shots binned in the hexagon
cell.
}
\examples{

# Construct hexagon shot chart (area/color varying)
hex_grid <- half_court_hex(cellsize = 20)
hex_grid <- merge_shot_data(hex_grid, dat_filt, hex = TRUE)
resized_hexes <- resize_hexes(hex_grid)
# Plot hexagon shot chart
breaks <- seq(0, max(hex_grid@data$FGP, na.rm = TRUE), length.out = 10)
plotclr <- c("#236e96", "#15b2d3", "#5abccc", "#ffbe42", "#ff7f00", "#f4543b", "#f4143b")
colcode <- plotclr[findInterval(resized_hexes@data$FGP, breaks)]
draw_halfcourt()
plot(resized_hexes, col = colcode, border = NA, add = TRUE)


}
