% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{team_ids}
\alias{team_ids}
\title{Get Team Identifiers for NBA}
\usage{
team_ids()
}
\value{
A data frame with one column representing team IDs and another column
  representing team abbreviations.
}
\description{
Wrapper for \code{\link[nba]{get_teaminfo}} to get NBA team identifiers.
}
\examples{
team_ids()

}
