% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{game_id}
\alias{game_id}
\title{Get Game Identifiers for NBA}
\usage{
game_id(year = "", game_number = "")
}
\arguments{
\item{year}{Year the season starts in YYYY format.}

\item{game_number}{Game number in the season.}
}
\examples{
game_id("2017", "1")
# check by visiting the url associated with the play-by-play
# of the game https://stats.nba.com/game/0021600001/playbyplay/
}
