% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_pbp2}
\alias{get_pbp2}
\title{Get Play-by-Play Data (V2)}
\usage{
get_pbp2(GameID = "", StartPeriod = "", EndPeriod = "", ...)
}
\arguments{
\item{GameID}{Game identifier.}

\item{StartPeriod}{TBD.}

\item{EndPeriod}{TBD.}

\item{...}{Arguments to code{\link[nba]{make_url}}.}
}
\value{
A data frame containing detailed play-by-play information for a single game.
}
\description{
Get Play-by-Play Data (V2)
}
