% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_shotchart}
\alias{get_shotchart}
\title{Get Shot Chart Data}
\usage{
get_shotchart(SeasonType = "", LeagueID = "", Season = "",
  PlayerID = "", TeamID = "", GameID = "", ContextMeasure = "",
  PlayerPosition = "", DateFrom = "", DateTo = "", GameSegment = "",
  LastNGames = "", Location = "", Month = "", OpponentTeamID = "",
  Outcome = "", SeasonSegment = "", VSConference = "", VSDivision = "",
  RookieYear = "", Period = "")
}
\value{
A data.frame containing shot chart data.
}
\description{
Get shot chart data for a player using the NBA API. See \code{\link[nba]{make_url}} for details on the arguments.
}
