% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{half_court_raster}
\alias{half_court_raster}
\title{Raster Half Court}
\usage{
half_court_raster()
}
\value{
SpatialPolygonsDataFrame representing the NBA regulation half court
  as a grid of rectangular polygons.
}
\description{
Construct a SpatialPolygonsDataFrame obejct out of a raster layer that
represents the NBA regulation half court.
}
\examples{
# create a half court raster SpatialPolygonsDataFrame object
sp_pol <- half_court_raster()

}
